backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<0.12
psycopg[binary]>=3.2,<4.0
dnspython>=2.6,<3.0
python-whois>=0.9,<1.0
tldextract>=5.1,<6.0
requests>=2.32,<3.0
django-storages>=1.14,<2.0
boto3>=1.35,<2.0
gunicorn>=23.0,<24.0